+++ lib/src/sort.dart
@@ -1,8 +1,8 @@
 import 'package:davi/src/sort_direction.dart';
 
 /// Describes the column sort.
-class ColumnSort {
-  ColumnSort({required this.columnIndex, required this.direction});
+class DaviSort {
+  DaviSort({required this.columnIndex, required this.direction});
 
   final int columnIndex;
   final DaviSortDirection direction;